+++ packages/frontend/src/header/Header.tsx
@@ -11,8 +11,10 @@ export const Header = () => {
         <StyledHeading size="medium">{" > it's hellfall"}</StyledHeading>
         <NavRow>
           <Navigation>
-            <Link to={'/'}>search</Link>, <Link to={'/card/random'}>random</Link>,{' '}
-            <Link to={'/draft'}>draft</Link>, <Link to={'/deck-builder'}>deck/cube builder</Link>,{' '}
+            <Link to={'/'}>search</Link>, <Link to={'/advanced'}>advanced</Link>,{' '}
+            {/* <Link to={'/syntax'}>syntax</Link>,  */}
+            <Link to={'/card/random'}>random</Link>, <Link to={'/draft'}>draft</Link>,{' '}
+            <Link to={'/deck-builder'}>deck/cube builder</Link>,{' '}
             <Link to={'/hellscubes'}>cube resources</Link>,{' '}
             <Link to={'https://discord.com/channels/631288872814247966/1237418389480407061'}>
               rules
