+++ packages/frontend/src/hellfall/TODO.md
@@ -139,3 +139,4 @@ Write a script to transform the database
 - add ability to have multiple showcases (Ian in Ixalan, Wizard Who Only Knows Fireball)
 - decide whether to have `-` invert result instead of inverting op
 - figure out if i need more specific extras handling if no sets are selected
+- move actualOp function to filterObjects?
