+++ packages/frontend/src/hellfall/filters/filterNumber.ts
@@ -1,22 +1,9 @@
 import { toNumber } from '../inputs/NumberSelector';
-import {
-  getActualOp,
-  invertOptionType,
-  looseOpType,
-  NOPRINT,
-  numFilter,
-  numStringFilter,
-  numStringListFilter,
-  opType,
-  textFilter,
-} from './types';
-import { isNumber } from '@hellfall/shared/utils/isInt.ts';
+import { invertOptionType, numFilter, numStringFilter, numStringListFilter, opType } from './types';
 
 export const filterNumber: numFilter = Object.assign(
-  function (this: numFilter, value1: number, operator: looseOpType, value2: number) {
-    const actualOp = getActualOp(this, operator);
-
-    switch (actualOp) {
+  (value1: number, operator: opType, value2: number) => {
+    switch (operator) {
       case '<':
         return value1 < value2;
       case '<=':
@@ -33,58 +20,45 @@ export const filterNumber: numFilter = Object.assign(
   },
   {
     invertOption: 'flip' as invertOptionType,
-    defaultOp: '=' as opType,
-    toSummary: (operator: looseOpType, value: number, invert?: boolean) =>
-      `${invert ? 'not' : ''} ${getActualOp(filterNumber, operator)} ${value}`,
+    toSummary: (operator: opType, value: number, invert?: boolean) =>
+      `${invert ? 'not' : ''} ${operator} ${value}`,
   }
 );
 
 export const filterNumberString: numStringFilter = Object.assign(
-  function (
-    this: numStringFilter,
-    value1: number | string | undefined,
-    operator: looseOpType,
-    value2: number | string | undefined
-  ) {
-    const actualOp = getActualOp(this, operator);
+  (value1: number | string | undefined, operator: opType, value2: number | string | undefined) => {
     const num1 = typeof value1 == 'string' ? toNumber(value1) : value1;
     const num2 = typeof value2 == 'string' ? toNumber(value2) : value2;
     if (num1 == undefined || num2 == undefined) {
       return false;
     }
-    return filterNumber(num1, actualOp, num2);
+    return filterNumber(num1, operator, num2);
   },
   {
     invertOption: 'flip' as invertOptionType,
-    defaultOp: '=' as opType,
-    toSummary: (operator: looseOpType, value: number | string | undefined, invert?: boolean) => {
+    toSummary: (operator: opType, value: number | string | undefined, invert?: boolean) => {
       const num = typeof value == 'string' ? toNumber(value) : value;
       if (num == undefined) {
         return `!The value must be a number (or convertible to one)`;
       }
-      return `${invert ? 'not' : ''} ${getActualOp(filterNumberString, operator)} ${num}`;
+      return `${invert ? 'not' : ''} ${operator} ${num}`;
     },
   }
 );
 export const filterNumberStringList: numStringListFilter = Object.assign(
-  function (
-    this: numStringListFilter,
+  (
     value1: (number | string | undefined)[],
-    operator: looseOpType,
+    operator: opType,
     value2: number | string | undefined
-  ) {
-    const actualOp = getActualOp(this, operator);
-    return value1.some(value => filterNumberString(value, actualOp, value2));
-  },
+  ) => value1.some(value => filterNumberString(value, operator, value2)),
   {
     invertOption: 'flip' as invertOptionType,
-    defaultOp: '=' as opType,
-    toSummary: (operator: looseOpType, value: number | string | undefined, invert?: boolean) => {
+    toSummary: (operator: opType, value: number | string | undefined, invert?: boolean) => {
       const num = typeof value == 'string' ? toNumber(value) : value;
       if (num == undefined) {
         return `!The value must be a number (or convertible to one)`;
       }
-      return `${invert ? 'not' : ''} ${getActualOp(filterNumberStringList, operator)} ${num}`;
+      return `${invert ? 'not' : ''} ${operator} ${num}`;
     },
   }
 );
