+++ packages/frontend/src/hellfall/filters/values/filterLayout.ts
@@ -1,14 +1,6 @@
-import { HCCard, HCLayout, HCLayoutGroup } from '@hellfall/shared/types';
-import {
-  cardStringFilter,
-  getActualOp,
-  invertOptionType,
-  looseOpType,
-  opToDont,
-  opToNot,
-  opType,
-  shareOp,
-} from '../types';
+import { HCLayout, HCLayoutGroup } from '@hellfall/shared/types';
+import { invertOptionType, opType, textListFilter } from '../types';
+import { shareOp, opToNot, opToDont } from '../filterUtils';
 export const toCardLayout: Record<string, HCLayout | HCLayout[]> = {
   normal: HCLayout.Normal,
   meldpart: HCLayout.MeldPart,
@@ -141,18 +133,12 @@ export const toFaceLayout: Record<
   planet: HCLayout.Station,
 };
 
-export const filterCardLayout: cardStringFilter = Object.assign(
-  function (this: cardStringFilter, value1: HCCard.Any, operator: looseOpType, value2: string) {
-    const actualOp = getActualOp(this, operator);
-    if (!(value2 in toCardLayout)) {
-      return false;
-    }
-    return shareOp(actualOp, value1.layout, toCardLayout[value2]);
-  },
+export const filterCardLayout: textListFilter = Object.assign(
+  (value1: string[], operator: opType, value2: string) =>
+    value2 in toCardLayout ? shareOp(operator, value1, toCardLayout[value2]) : false,
   {
     invertOption: 'flip' as invertOptionType,
-    defaultOp: '=' as opType,
-    toSummary: (operator: looseOpType, value: string) => {
+    toSummary: (operator: opType, value: string) => {
       if (!(value in toCardLayout)) {
         return `!Unknown card layout "${value}"`;
       }
@@ -169,22 +155,12 @@ export const filterCardLayout: cardStringFilter = Object.assign(
     },
   }
 );
-export const filterFaceLayout: cardStringFilter = Object.assign(
-  function (this: cardStringFilter, value1: HCCard.Any, operator: looseOpType, value2: string) {
-    const actualOp = getActualOp(this, operator);
-    if (!(value2 in toFaceLayout)) {
-      return false;
-    }
-    return shareOp(
-      actualOp,
-      value1.toFaces().map(e => e.layout),
-      toFaceLayout[value2]
-    );
-  },
+export const filterFaceLayout: textListFilter = Object.assign(
+  (value1: string[], operator: opType, value2: string) =>
+    value2 in toFaceLayout ? shareOp(operator, value1, toFaceLayout[value2]) : false,
   {
     invertOption: 'flip' as invertOptionType,
-    defaultOp: '=' as opType,
-    toSummary: (operator: looseOpType, value: string) => {
+    toSummary: (operator: opType, value: string) => {
       if (!(value in toFaceLayout)) {
         return `!Unknown face layout "${value}"`;
       }
@@ -204,15 +180,12 @@ export const filterFaceLayout: cardStringFilter = Object.assign(
     },
   }
 );
-export const filterAnyLayout: cardStringFilter = Object.assign(
-  function (this: cardStringFilter, value1: HCCard.Any, operator: looseOpType, value2: string) {
-    const actualOp = getActualOp(this, operator);
-    return filterCardLayout(value1, actualOp, value2) || filterFaceLayout(value1, actualOp, value2);
-  },
+export const filterAnyLayout: textListFilter = Object.assign(
+  (value1: string[], operator: opType, value2: string) =>
+    filterCardLayout(value1, operator, value2) || filterFaceLayout(value1, operator, value2),
   {
     invertOption: 'flip' as invertOptionType,
-    defaultOp: '=' as opType,
-    toSummary: (operator: looseOpType, value: string) => {
+    toSummary: (operator: opType, value: string) => {
       const cardSum = filterCardLayout.toSummary(operator, value);
       const faceSum = filterFaceLayout.toSummary(operator, value);
       return cardSum[0] != '!'
