+++ packages/frontend/src/hellfall/filters/values/filterLegality.ts
@@ -1,69 +1,37 @@
 import { formatList, HCFormat, HCLegalitiesField } from '@hellfall/shared/types';
-import {
-  funcOp,
-  legalFilter,
-  looseOpType,
-  opType,
-  textFilter,
-  getActualOp,
-  opToNot,
-  NOPRINT,
-  invertOptionType,
-} from '../types';
+import { legalFilter, opType, invertOptionType } from '../types';
+import { funcOp, opToNot } from '../filterUtils';
 
 export const filterLegal: legalFilter = Object.assign(
-  function (this: legalFilter, value1: HCLegalitiesField, operator: looseOpType, value2: string) {
-    const actualOp = getActualOp(this, operator);
-    return funcOp(actualOp, format => value1[format as HCFormat] == 'legal', value2);
-  },
+  (value1: HCLegalitiesField, operator: opType, value2: string) =>
+    funcOp(operator, format => value1[format as HCFormat] == 'legal', value2),
   {
     invertOption: 'flip' as invertOptionType,
-    defaultOp: '=' as opType,
-    toSummary: (operator: looseOpType, value: string) =>
+    toSummary: (operator: opType, value: string) =>
       formatList.includes(value)
         ? `it's ${opToNot(operator)} legal in ${value}`
         : `!Unknown format "${value}"`,
   }
 );
 export const filterBanned: legalFilter = Object.assign(
-  function (this: legalFilter, value1: HCLegalitiesField, operator: looseOpType, value2: string) {
-    const actualOp = getActualOp(this, operator);
-    return funcOp(actualOp, format => value1[format as HCFormat] == 'banned', value2);
-  },
+  (value1: HCLegalitiesField, operator: opType, value2: string) =>
+    funcOp(operator, format => value1[format as HCFormat] == 'banned', value2),
   {
     invertOption: 'flip' as invertOptionType,
-    defaultOp: '=' as opType,
-    toSummary: (operator: looseOpType, value: string) =>
+    toSummary: (operator: opType, value: string) =>
       formatList.includes(value)
         ? `it's ${opToNot(operator)} banned in ${value}`
         : `!Unknown format "${value}"`,
   }
 );
 export const filterNotLegal: legalFilter = Object.assign(
-  function (this: legalFilter, value1: HCLegalitiesField, operator: looseOpType, value2: string) {
-    const actualOp = getActualOp(this, operator);
-    return funcOp(actualOp, format => value1[format as HCFormat] == 'not_legal', value2);
-  },
+  (value1: HCLegalitiesField, operator: opType, value2: string) =>
+    funcOp(operator, format => value1[format as HCFormat] == 'not_legal', value2),
   {
     invertOption: 'flip' as invertOptionType,
-    defaultOp: '=' as opType,
-    toSummary: (operator: looseOpType, value: string) =>
+    toSummary: (operator: opType, value: string) =>
       formatList.includes(value)
         ? `it's ${opToNot(operator)} notlegal in ${value}`
         : `!Unknown format "${value}"`,
-  }
-);
-/**
- * @deprecated Only used in old search logic.
- */
-export const filterLegality: textFilter = Object.assign(
-  function (this: textFilter, value1: string, operator: looseOpType, value2: string) {
-    const actualOp = getActualOp(this, operator);
-    return funcOp(actualOp, legality => legality == value2, value1);
-  },
-  {
-    invertOption: 'flip' as invertOptionType,
-    defaultOp: '=' as opType,
-    toSummary: (operator: looseOpType, value: string) => NOPRINT,
   }
 );
